Inclusao Modals e outros pequenos ajustes

+++ salareunioesangular/src/app/room-list/room-list.component.ts
@@ -1,8 +1,11 @@
-import { Component, Injectable, OnInit } from '@angular/core';
+import { Component, OnInit } from '@angular/core';
+import { MatDialog } from '@angular/material/dialog';
 import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { Room } from '../room';
 import { RoomService } from '../room.service';
+import { Alerta } from '../shared/alerta/alerta';
+import { AlertadeleteComponent } from '../shared/alertadelete/alertadelete.component';
 
 @Component({
   selector: 'app-room-list',
@@ -13,7 +16,7 @@ export class RoomListComponent implements OnInit {
 
   rooms!: Observable<Room[]>
 
-  constructor(private roomService: RoomService,
+  constructor(private roomService: RoomService, public dialog: MatDialog,
     private router: Router) {}
 
   ngOnInit() {
@@ -25,13 +28,37 @@ export class RoomListComponent implements OnInit {
   }
 
   deleteRoom(id: number) {
-    this.roomService.deleteRoom(id)
-      .subscribe(
-        (        data: any) => {
-          console.log(data);
-          this.reloadData();
-        },
-        (        error: any) => console.log(error));
+
+      const config = {
+        data: {
+          btnSucesso: 'Confirm',
+          btnCancelar: 'Back to Room List',
+          corBtnCancelar: 'primary',
+          possuirBtnFechar: true
+        } as Alerta
+      };
+      const dialogRef = this.dialog.open(AlertadeleteComponent, config);
+      dialogRef.afterClosed().subscribe((opcao: boolean) => {
+        if (opcao) {
+          this.roomService.deleteRoom(id)
+          .subscribe(
+            (        data: any) => {
+              console.log(data);
+              this.reloadData();
+            },
+            () => {
+              const config = {
+                data: {
+                  titulo: 'Error!',
+                  descricao: 'It was not possible save your register, please try again',
+                  corBtnSucesso: 'warn',
+                  btnSucesso: 'Close'
+                } as Alerta
+              };
+              this.dialog.open(AlertadeleteComponent, config);
+            })
+        } 
+      });
   }
 
   roomDetails(id: number){
